# llmproxy/ui/openapi.yaml
openapi: 3.0.3
info:
  title: llmproxy API
  description: |
    自托管 LLM 转发网关。下游 OpenAI 兼容；管理与自助接口在 `/v1/_admin` / `/v1/_me`。

    鉴权：下游用 `Authorization: Bearer <api_key 或用户 token>`；
    管理接口用 `Authorization: Bearer <admin_token>`。
  version: "2.3.0"
servers:
  - url: http://127.0.0.1:8787
    description: 本机默认
...
paths:
  /v1/chat/completions:
    post:
      tags: [openai]
      summary: 对话补全
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatRequest' }
      responses:
        '200':
          description: 成功（非流式 JSON 或流式 SSE）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatResponse' }
            text/event-stream:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { description: 配额不足 }
        '429': { description: 限流 }
        '502': { description: 上游不可用 }
  /v1/models:
    get:
      tags: [openai]
      summary: 可用模型列表
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: 模型列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  object: { type: string, example: list }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Model' }
  /v1/_me:
    get:
      tags: [me]
      summary: 身份、上游与累计用量
      security: [{bearerAuth: []}]
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 下游 API key 或用户 token
    adminAuth:
      type: http
      scheme: bearer
      description: admin_token
  responses:
    Unauthorized:
      description: 未鉴权
  schemas:
    ChatRequest:
      type: object
      required: [model, messages]
      properties:
        model: { type: string }
        messages:
          type: array
          items:
            type: object
            properties:
              role: { type: string }
              content: { type: string }
        stream: { type: boolean }
        temperature: { type: number }
    ChatResponse:
      type: object
      properties:
        id: { type: string }
        object: { type: string }
        choices: { type: array, items: { type: object } }
        usage: { type: object }
    Model:
      type: object
      properties:
        id: { type: string }
        object: { type: string }
...